third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristic/gen-service-is-removed.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Service is removed before getCharacteristic call. ' +
  'Reject with InvalidStateError.';
let service, fake_service, fake_peripheral;

bluetooth_test(() => getHealthThermometerService()
    .then(_ => ({service, fake_service, fake_peripheral} = _))
    .then(() => fake_service.remove())
    .then(() => fake_peripheral.simulateGATTServicesChanged())
    .then(() => assert_promise_rejects_with_message(
      service.getCharacteristic('measurement_interval'),
      new DOMException('GATT Service no longer exists.',
                       'InvalidStateError'),
      'Service got removed.')),
    test_desc);

</script>
